+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -83,5 +83,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-05-15T06:43:27.397Z")
+    updated: Optional[datetime] = parse_datetime("2025-05-16T08:02:19.729Z")
     r"""The date and time the resource was last updated"""