+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JV9AY7KPH4GJYBG1XFT1MZX0"
+        "01JVC1VC0F2RH0ZEYT7GP9KW3P"
     )
     r"""The id of the resource"""