+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -93,5 +93,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-05-15T06:43:30.109Z")
+    created: Optional[datetime] = parse_datetime("2025-05-16T08:02:22.393Z")
     r"""The date and time the resource was created"""