+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JV9AY7KRWTMJDFP57FPS7JGA"
+        "01JVC1VC0G92T9SV4GY71SE880"
     )
     r"""The id of the resource"""